Attempt to use Vite instead of Webpack
It seems not great for this purpose. Development is assumed to be on an
entirely self-contained library or project. Development is only accessed
via an index.html - this doesn't allow for dependency SCSS being pulled
from another project (even just to be able to test via index.html).

Our situation relies on testing being done _within_ the target
environment, the Vite system doesn't seem well set up for plugin or
peerDependency development. It can probably do it, but might not be the
best tool for the job.

E.g. directly defining scss files to process doesn't seem to be a thing.
They can be listed in build options, but this precludes dev time
functionality and we may as well just use Rollup directly. Plus it
leaves nonsense js files (from transpilation) in the output.

Configuration of sass is also quite direct, controlling the outfile
seems difficult.

Importing dependencies from admin is troublesome as they often include
jsx syntax and Vite does not expect this (causing Rollup to error). I do
not see documentation on being able to set .js files to be parsed as
.jsx

The tool is probably great if one sets up the expected project
structure, it is very opinionated and stepping outside this expectation
results in a lot of manual configuration to the point one may as well
just use the underlying tools directly. The Vite layer on top is
essentially bypassed, so this tool is not great for our purpose.

Building full webapp projects - probably great.
Building pure JS libraries - probably alright.
Building a sub-product for an existing webapp ecosystem - terrible.