Try Parcel instead of Webpack
Another "Zero Conf" bundler. Except this one isn't a bundler. And again
the zero configuration claim means "build a webapp or simple library,
otherwise be prepared for myriad of configuration".

I think I'll quit before I get too deep in it.
It looks like the tool can be configured to do what we want, but the
documentation focusses on capability/ethos, not technical specifics,
which makes it very hard to follow.