Apply shortcodes + content alterations from editor
When inputting text one can open an editor to apply, alter, or remove
shortcodes to the input element. Previously although the editor has
opened, the selected actions within it have not correctly applied.

The content field on the editor poses the most interesting challenge as
it updates the content of the input, not just apply (or remove) the
dentoation of a shortcode.