FIX show configuration errors instead of no output
PHP is a permissive loosely typed language that will happily turn
initialise a variable on first access instead of erroring. This lets
programmers check `$val = $undefined`, `$val = $array['non-value']`,
and `$val = $validObject->invalidProperty` and get have the value of
`$val` be `null` in all examples.

This in turn means that `ViewableData::obj()` will then cast `null`
into a value object (DBText by default) as it is a non-object.

This is confusing to the end user of DataObjectPropertyDisplay
(myself in the first instance) as the expected output of our
configured instance is absent, i.e. `''` (empty string) where we
would otherwise expect a value to be shown.