FIX incorrect DB field exists() values
If one implements a getter method for use in templates that uses the
value from e.g. Boolean (Framework's DBBoolean), the value is returned,
rather than the casted value-object (as would be fetched via the
template). This results in the output of 0 being false for exists calls
under normal circumstances.

However when replacing the default cast (DBText) with our new
ShortcodeText that has the shortcode parser is enabled, the value 0
takes a trip through the shortcode parser that always returns a string,
'0'... which in turn would normally be false in PHP, but the default
template logic counts this as valid, returning true for exists calls.

The result is that custom getters with non-string values that should be
false suddenly return true, messing up template output and not being
properly backwards compatiable/a safe drop in replacement.

Functionality must be kept in parity with the field type it replaces in
order to be achieve the modules goal.