FIX plain text field convert nl to BR in templates
DBString (from framework) performs a PHP nl2br() on its value before
outputting it via `forTemplate`. Our new override that is supposed to be
a drop-in replacement neglected to do this, resulting in plain text
data held in the CMS (e.g. type 'Text') printing all on a single line.
Fine in simple cases, but not for e.g. TextArea where an author would
expect new lines to be preserved.